+++ next.config.js
@@ -1,5 +1,3 @@
-const STORE_ID = process.env.STORE_ID;
-
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
@@ -8,14 +6,6 @@ const nextConfig = {
   },
   experimental: {
     nextScriptWorkers: true,
-  },
-  async rewrites() {
-    return [
-      {
-        source: '/vtex/api/:path*',
-        destination: `https://${STORE_ID}.myvtex.com/api/:path*`,
-      },
-    ];
   },
   webpack: (config, { isServer, dev }) => {
     if (!isServer && !dev && config.optimization?.splitChunks) {
