+++ package.json
@@ -44,6 +44,7 @@
     "@hookform/resolvers": "^3.1.0",
     "@next/font": "13.1.1",
     "@splidejs/react-splide": "^0.7.12",
+    "@types/http-proxy": "^1.17.11",
     "@types/node": "18.11.18",
     "@types/promise-queue": "^2.2.0",
     "@types/react": "18.0.26",
@@ -57,6 +58,7 @@
     "graphql": "^16.6.0",
     "graphql-request": "^5.2.0",
     "graphql-yoga": "^3.8.0",
+    "http-proxy": "^1.18.1",
     "jotai": "^2.0.4",
     "localforage": "^1.10.0",
     "localforage-driver-memory": "^1.0.5",
