+++ src/components/molecules/Auth/LoggedPopup/LoggedPopup.tsx
@@ -10,7 +10,7 @@ import {
   PopoverTrigger,
   useToast,
 } from '@chakra-ui/react';
-import Auth from '@services/auth/Auth';
+import Auth from '@services/auth';
 import NextLink from 'next/link';
 import { useRouter } from 'next/router';
 import { AiOutlineUser } from 'react-icons/ai';
