+++ src/components/organisms/Login/AccessKey/EmailForm.tsx
@@ -7,7 +7,7 @@ import {
   chakra,
 } from '@chakra-ui/react';
 import { yupResolver } from '@hookform/resolvers/yup';
-import Auth from '@services/auth/Auth';
+import Auth from '@services/auth';
 import { useForm } from 'react-hook-form';
 import { useMutation } from 'react-query';
 import { useLoginContext } from '../Login';
