+++ src/pages/api/auth/logout.ts
@@ -1,4 +1,4 @@
-import VtexId from '@services/vtex-id/VtexId';
+import VtexId from '@services/vtex-id';
 import { deleteCookie } from 'cookies-next';
 import { NextApiRequest, NextApiResponse } from 'next';
 import { ONE_DAY } from 'src/sdk/constants';
