+++ src/services/safedata/index.ts
@@ -4,7 +4,7 @@ import { User } from './types';
 class SafeData {
   async getUserData() {
     const { data } = await axios.get<User[]>(
-      '/vtex/api/io/safedata/CL/search?_fields=email,document,firstName,lastName,phone,birthDate,userId',
+      '/api/vtex/io/safedata/CL/search?_fields=email,document,firstName,lastName,phone,birthDate,userId',
     );
 
     return data;
@@ -12,7 +12,7 @@ class SafeData {
 
   async updateUserData(userId: string, newData: Partial<User>) {
     const { data } = await axios.patch<User>(
-      `/vtex/api/io/safedata/CL/documents/${userId}`,
+      `/api/vtex/io/safedata/CL/documents/${userId}`,
       newData,
     );
